fix: move api key sync to a listener
